+++ HISTORY.rst
@@ -4,7 +4,9 @@ History
 2.2.1 (UNRELEASED)
 ------------------
 
+* Add ``--skip-constraints`` option.
 * Fix bootstrapping for autoresolve case with missing output files.
+
 
 2.2.0 (2020-01-22)
 ------------------